+++ Plugins/D3D12/D3D12CommandBuffer.cpp
@@ -318,7 +318,7 @@ void cmd::set_index_buffer(GPUCommandEncoderHandle cmdbuffer, GPUBufferHandle bu
     // create the index buffer view
     D3D12_INDEX_BUFFER_VIEW index_buffer_view = {};
     index_buffer_view.BufferLocation          = buf.buffer->GetGPUVirtualAddress() + offset;
-    index_buffer_view.SizeInBytes             = size == 0ull ? buf.size() : size;
+    index_buffer_view.SizeInBytes             = static_cast<uint>(size == 0ull ? buf.size() : size);
     index_buffer_view.Format                  = d3d12enum(format);
 
     // set the index buffer
@@ -336,7 +336,7 @@ void cmd::set_vertex_buffer(GPUCommandEncoderHandle cmdbuffer, GPUIndex32 slot, 
     // create vertex buffer view
     D3D12_VERTEX_BUFFER_VIEW vertex_buffer_view = {};
     vertex_buffer_view.BufferLocation           = buf.buffer->GetGPUVirtualAddress() + offset;
-    vertex_buffer_view.SizeInBytes              = size == 0ull ? buf.size() : size;
+    vertex_buffer_view.SizeInBytes              = static_cast<uint>(size == 0ull ? buf.size() : size);
     vertex_buffer_view.StrideInBytes            = pip->vertex_buffer_strides.at(slot);
 
     // set the vertex buffer at the specified slot
@@ -694,7 +694,7 @@ void cmd::buffer_barrier(GPUCommandEncoderHandle cmdbuffer, GPUBufferBarriers ba
 
     D3D12_BARRIER_GROUP barrier_group = {};
     barrier_group.Type                = D3D12_BARRIER_TYPE_BUFFER;
-    barrier_group.NumBarriers         = barriers.size();
+    barrier_group.NumBarriers         = static_cast<uint>(barriers.size());
     barrier_group.pBufferBarriers     = d3d12_barriers.data();
 
     // issue the resource barriers
@@ -735,7 +735,7 @@ void cmd::texture_barrier(GPUCommandEncoderHandle cmdbuffer, GPUTextureBarriers 
 
     D3D12_BARRIER_GROUP barrier_group = {};
     barrier_group.Type                = D3D12_BARRIER_TYPE_TEXTURE;
-    barrier_group.NumBarriers         = barriers.size();
+    barrier_group.NumBarriers         = static_cast<uint>(barriers.size());
     barrier_group.pTextureBarriers    = d3d12_barriers.data();
 
     // issue the resource barriers
